+++ src/spooq2/transformer/mapper.py
@@ -106,15 +106,17 @@ class Mapper(Transformer):
     Attention: Decimal is NOT SUPPORTED by Hive! Please use Double instead!
     """
 
-    def __init__(self, mapping, ignore_missing_columns=False, mode="replace"):
+    def __init__(self, mapping, ignore_missing_columns=False, ignore_ambiguous_columns=False, mode="replace"):
         super(Mapper, self).__init__()
         self.mapping = mapping
         self.ignore_missing_columns = ignore_missing_columns
+        self.ignore_ambiguous_columns = ignore_ambiguous_columns
         self.mode = mode
 
     def transform(self, input_df):
         self.logger.info("Generating SQL Select-Expression for Mapping...")
-        self.logger.debug("Input Schema/Mapping: {mp}".format(mp=str(self.mapping)))
+        self.logger.debug("Input Schema/Mapping:")
+        self.logger.debug("\n" + "\n".join(["\t".join(map(str, mapping_line)) for mapping_line in self.mapping]))
 
         input_columns = input_df.columns
         select_expressions = []
@@ -126,6 +128,8 @@ class Mapper(Transformer):
             )
 
             source_column = self._get_spark_column(source_column, name, input_df)
+            if source_column is None:
+                continue
             data_type, data_type_is_spark_builtin = self._get_spark_data_type(data_type)
             select_expression = self._get_select_expression(name, source_column, data_type, data_type_is_spark_builtin)
 
@@ -137,7 +141,7 @@ class Mapper(Transformer):
                 select_expressions.append(select_expression)
 
         self.logger.info("SQL Select-Expression for new mapping generated!")
-        self.logger.debug("SQL Select-Expressions for new mapping: " + str(select_expressions))
+        self.logger.debug("SQL Select-Expressions for new mapping:\n" + "\n".join(str(select_expressions).split(",")))
         self.logger.debug("SQL WithColumn-Expressions for new mapping: " + str(with_column_expressions))
         if self.mode == "prepend":
             df_to_return = input_df.select(select_expressions + ["*"])
@@ -168,7 +172,11 @@ class Mapper(Transformer):
 
         except AnalysisException as e:
             if isinstance(source_column, str) and self.ignore_missing_columns:
+                self.logger.warn(f"Missing column ({str(source_column)}) replaced with NULL (via ignore_missing_columns=True): {e.desc}")
                 source_column = F.lit(None)
+            elif e.desc.startswith("Ambiguous reference to fields") and self.ignore_ambiguous_columns:
+                self.logger.warn(f"Exception ignored (via ignore_ambiguous_columns=True) for column \"{str(source_column)}\": {e.desc}")
+                return None
             else:
                 self.logger.exception(
                     "Column: \"{}\" cannot be resolved ".format(str(source_column)) +