+++ tests/unit/transformer/test_flattener.py
@@ -11,6 +11,16 @@ def flattener():
     return Flattener(pretty_names=False)
 
 
+def assert_mapping_equality(mapping_1, mapping_2, spark):
+    if mapping_1 == mapping_2:
+        return True
+    else:  # for easier debugging
+        assert_df_equality(
+            spark.createDataFrame(mapping_1, ["name", "source", "type"]),
+            spark.createDataFrame(mapping_2, ["name", "source", "type"])
+        )
+
+
 class TestBasicAttributes:
     """Mapper for Flattening DataFrames"""
 
@@ -67,7 +77,7 @@ class TestDataFrameContainingArrays:
         )])
         output_df = flattener.transform(input_df)
         expected_output_df = spark_session.createDataFrame([
-            (4789, ), (4790, ), (4791, )], schema=["array_val_exploded"])
+            (4789, ), (4790, ), (4791, )], schema=["array_val"])
 
         assert_df_equality(expected_output_df, output_df)
 
@@ -81,7 +91,7 @@ class TestDataFrameContainingArrays:
             (datetime.datetime(2021, 1, 14, 8, 10, 14), 4789),
             (datetime.datetime(2021, 1, 14, 8, 10, 14), 4790),
             (datetime.datetime(2021, 1, 14, 8, 10, 14), 4791)],
-            schema=["timestamp_val", "array_val_exploded"])
+            schema=["timestamp_val", "array_val"])
 
         assert_df_equality(expected_output_df, output_df)
 
@@ -104,7 +114,7 @@ class TestDataFrameContainingArrays:
             (4791, "Are"),
             (4791, "You"),
             (4791, "?")],
-            schema=["array_val_1_exploded", "array_val_2_exploded"])
+            schema=["array_val_1", "array_val_2"])
 
         assert_df_equality(expected_output_df, output_df)
 
@@ -128,7 +138,7 @@ class TestDataFrameContainingArrays:
             (43.102, 4791, "Are"),
             (43.102, 4791, "You"),
             (43.102, 4791, "?")],
-            schema=["double_val", "array_val_1_exploded", "array_val_2_exploded"])
+            schema=["double_val", "array_val_1", "array_val_2"])
 
         assert_df_equality(expected_output_df, output_df)
 
@@ -147,7 +157,7 @@ class TestDataFrameContainingArrays:
             ("How are you?", "127"),
             ("How are you?", "53"),
             ("How are you?", "90")],
-            schema=["string_val", "array_val_exploded_exploded"])
+            schema=["string_val", "array_val"])
 
         assert_df_equality(expected_output_df, output_df)
 
@@ -161,7 +171,7 @@ class TestDataFrameContainingArrays:
             ("How are you?", 4789),
             ("How are you?", 4790),
             ("How are you?", 4791)],
-            schema=["struct_val_string_val", "struct_val_array_val_exploded"])
+            schema=["struct_val_string_val", "struct_val_array_val"])
 
         assert_df_equality(expected_output_df, output_df)
 
@@ -179,8 +189,8 @@ class TestDataFrameContainingArrays:
         expected_output_df = spark_session.createDataFrame([
             (43.102, 4789, "Hello Darkness", datetime.date(2021, 1, 14)),
             (43.102, 4790, "My Old Friend", datetime.date(2021, 1, 15))],
-            schema=["double_val", "array_val_exploded_int_val",
-                    "array_val_exploded_string_val", "array_val_exploded_date_val"])
+            schema=["double_val", "array_val_int_val",
+                    "array_val_string_val", "array_val_date_val"])
 
         assert_df_equality(expected_output_df, output_df)
 
@@ -230,8 +240,10 @@ class TestDataFrameContainingStructs:
 
 
 class TestComplexRecipes:
-    @pytest.fixture()
+
+    @pytest.fixture(scope="class")
     def input_df(self, spark_session):
+        """Taken from https://opensource.adobe.com/Spry/samples/data_region/JSONDataSetSample.html"""
         return spark_session.createDataFrame([Row(
             batters=Row(
                 batter=[Row(id="1001", type="Regular"),
@@ -251,9 +263,9 @@ class TestComplexRecipes:
             type="donut",
         )])
 
-    @pytest.fixture()
-    def expected_output_df(self, spark_session, flattener):
-        return spark_session.createDataFrame([
+    @pytest.fixture(scope="class")
+    def expected_output_data(self):
+        return [
             ("0001", "Cake", 0.55, "donut", "1001", "Regular",      "5001", "None"                     ),
             ("0001", "Cake", 0.55, "donut", "1001", "Regular",      "5002", "Glazed"                   ),
             ("0001", "Cake", 0.55, "donut", "1001", "Regular",      "5005", "Sugar"                    ),
@@ -281,14 +293,30 @@ class TestComplexRecipes:
             ("0001", "Cake", 0.55, "donut", "1004", "Devil's Food", "5007", "Powdered Sugar"           ),
             ("0001", "Cake", 0.55, "donut", "1004", "Devil's Food", "5006", "Chocolate with Sprinkles" ),
             ("0001", "Cake", 0.55, "donut", "1004", "Devil's Food", "5003", "Chocolate"                ),
-            ("0001", "Cake", 0.55, "donut", "1004", "Devil's Food", "5004", "Maple"                    )],
-            schema=["id", "name", "ppu", "type", "batters_batter_exploded_id", "batters_batter_exploded_type",
-                    "topping_exploded_id", "topping_exploded_type", ]
-        )
+            ("0001", "Cake", 0.55, "donut", "1004", "Devil's Food", "5004", "Maple"                    )]
 
-    def test_donut(self, input_df, expected_output_df):
+    @pytest.fixture(scope="class")
+    def expected_output_df(self, expected_output_data, spark_session):
+        return spark_session.createDataFrame(
+            expected_output_data,
+            schema=["id", "name", "ppu", "type", "batters_batter_id", "batters_batter_type",
+                    "topping_id", "topping_type"])
+
+    @pytest.fixture(scope="class")
+    def expected_output_df_pretty(self, expected_output_data, spark_session):
+        return spark_session.createDataFrame(
+            expected_output_data,
+            schema=["id", "name", "ppu", "type", "batter_id", "batter_type",
+                    "topping_id", "topping_type"])
+
+    def test_donut(self, input_df, expected_output_df, flattener):
         output_df = flattener.transform(input_df)
         assert_df_equality(expected_output_df, output_df)
+
+    def test_pretty_donut(self, input_df, expected_output_df_pretty, flattener):
+        flattener.pretty_names = True
+        output_df = flattener.transform(input_df)
+        assert_df_equality(expected_output_df_pretty, output_df)
 
 
 class TestPrettyColumnNames:
@@ -297,18 +325,115 @@ class TestPrettyColumnNames:
     def flattener(self):
         return Flattener(pretty_names=True)
 
-    def test_simple_renames(self, flattener):
-        input_mapping = [
-            ("double_val",                      "double_val",                       "double"),
-            ("array_val_exploded_int_val",      "array_val_exploded_int_val",       "int"),
-            ("array_val_exploded_string_val",   "array_val_exploded_string_val",    "double"),
-            ("array_val_exploded_date_val",     "array_val_exploded_date_val",      "date"),
-        ]
-        expected_output_mapping = [
-            ("double_val",   "double_val",                       "double"),
-            ("int_val",      "array_val_exploded_int_val",       "int"),
-            ("string_val",   "array_val_exploded_string_val",    "double"),
-            ("date_val",     "array_val_exploded_date_val",      "date"),
-        ]
-        output_mapping = flattener._prettify_column_names(input_mapping)
-        assert output_mapping == expected_output_mapping
+    def test_simple_renames(self, flattener, spark_session):
+        input_df = spark_session.createDataFrame([Row(
+            struct_val=Row(int_val=4789, string_val="Hello World")
+        )])
+        expected_output_df = spark_session.createDataFrame([
+            (4789, "Hello World")], schema=["int_val", "string_val"])
+        output_df = flattener.transform(input_df)
+        assert_df_equality(output_df, expected_output_df)
+
+    def test_duplicated_column_names(self, flattener, spark_session):
+        input_df = spark_session.createDataFrame([Row(
+            struct_val=Row(int_val=4789, string_val="Hello World"),
+            struct_val_2=Row(int_val=4790, string_val="How are you?")
+        )])
+        expected_output_df = spark_session.createDataFrame([
+            (4789, "Hello World", 4790, "How are you?")], schema=["int_val", "string_val", "struct_val_2_int_val", "struct_val_2_string_val"])
+        output_df = flattener.transform(input_df)
+        assert_df_equality(output_df, expected_output_df)
+
+    def test_nested_struct_attributes(self, flattener, spark_session):
+        input_df = spark_session.createDataFrame([Row(
+            struct_val_1=Row(
+                struct_val_2=Row(
+                    struct_val_3=Row(
+                        struct_val_4=Row(int_val=4789),
+                        int_val=4790),
+                    string_val="Hello"),
+                double_val=43.12),
+            timestamp_val=datetime.datetime(2021, 1, 1, 12, 30, 15)
+        )])
+        output_df = flattener.transform(input_df)
+        expected_output_df = spark_session.createDataFrame(
+            [(4789, 4790, "Hello", 43.12, datetime.datetime(2021, 1, 1, 12, 30, 15))],
+            schema=["int_val", "struct_val_3_int_val", "string_val", "double_val", "timestamp_val"]
+        )
+        assert_df_equality(expected_output_df, output_df)
+
+    def test_multiple_arrays_with_other_columns(self, flattener, spark_session):
+        input_df = spark_session.createDataFrame([Row(
+            array_val_1=[4789, 4790, 4791],
+            array_val_2=["How", "Are", "You", "?"],
+            double_val=43.102
+        )])
+        output_df = flattener.transform(input_df)
+        expected_output_df = spark_session.createDataFrame([
+            (43.102, 4789, "How"),
+            (43.102, 4789, "Are"),
+            (43.102, 4789, "You"),
+            (43.102, 4789, "?"),
+            (43.102, 4790, "How"),
+            (43.102, 4790, "Are"),
+            (43.102, 4790, "You"),
+            (43.102, 4790, "?"),
+            (43.102, 4791, "How"),
+            (43.102, 4791, "Are"),
+            (43.102, 4791, "You"),
+            (43.102, 4791, "?")],
+            schema=["double_val", "array_val_1", "array_val_2"])
+
+        assert_df_equality(expected_output_df, output_df)
+
+    def test_array_nested_in_array(self, spark_session, flattener):
+        input_df = spark_session.createDataFrame([Row(
+            array_val=[["Here's", "My", "Number", ":"], [555, 127, 53, 90]],
+            string_val="How are you?"
+        )])
+        output_df = flattener.transform(input_df)
+        expected_output_df = spark_session.createDataFrame([
+            ("How are you?", "Here's"),
+            ("How are you?", "My"),
+            ("How are you?", "Number"),
+            ("How are you?", ":"),
+            ("How are you?", "555"),
+            ("How are you?", "127"),
+            ("How are you?", "53"),
+            ("How are you?", "90")],
+            schema=["string_val", "array_val"])
+
+        assert_df_equality(expected_output_df, output_df)
+
+    def test_array_nested_in_struct(self, spark_session, flattener):
+        input_df = spark_session.createDataFrame([Row(
+            struct_val=Row(array_val=[4789, 4790, 4791],
+                           string_val="How are you?")
+        )])
+        output_df = flattener.transform(input_df)
+        expected_output_df = spark_session.createDataFrame([
+            ("How are you?", 4789),
+            ("How are you?", 4790),
+            ("How are you?", 4791)],
+            schema=["string_val", "array_val"])
+
+
+        assert_df_equality(expected_output_df, output_df)
+
+    def test_struct_nested_in_array(self, spark_session, flattener):
+        input_df = spark_session.createDataFrame([Row(
+            array_val=[Row(int_val=4789,
+                           string_val="Hello Darkness",
+                           date_val=datetime.date(2021, 1, 14)),
+                       Row(int_val=4790,
+                           string_val="My Old Friend",
+                           date_val=datetime.date(2021, 1, 15))],
+            double_val=43.102
+        )])
+        output_df = flattener.transform(input_df)
+        expected_output_df = spark_session.createDataFrame([
+            (43.102, 4789, "Hello Darkness", datetime.date(2021, 1, 14)),
+            (43.102, 4790, "My Old Friend", datetime.date(2021, 1, 15))],
+            schema=["double_val", "int_val", "string_val", "date_val"])
+
+        assert_df_equality(expected_output_df, output_df)